// socialMedia-V2-main/src/common/enums/notification.enum.ts
export enum NotificationTypeEnum {
  FRIEND_REQUEST = "friend_request",
  FRIEND_ACCEPTED = "friend_accepted",
  COMMENT = "comment",
  COMMENT_REPLY = "comment_reply",
  COMMENT_MENTION = "comment_mention",
  POST_MENTION = "post_mention",
  POST_SHARE = "post_share",
  POST_REACTION = "post_reaction",
  COMMENT_REACTION = "comment_reaction",
  MESSAGE = "message",
}

export enum DevicePlatformEnum {
  WEB = "web",
  ANDROID = "android",
  IOS = "ios",
}